requirements.txt
# Core dependencies
langgraph==0.2.14
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Email Service
sib-api-v3-sdk==7.6.0
email-validator==2.1.0

# API and HTTP
requests==2.31.0
httpx==0.27.0

# Data handling
pydantic>=2.7.4
pandas>=2.1.4
numpy>=1.26.0

# Environment and configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging and monitoring
structlog==23.2.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
jinja2==3.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development
black==23.12.1
isort==5.13.2
flake8==6.1.0